% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package.R
\name{macument}
\alias{macument}
\title{Use Markdown and Roxygen to document a package}
\usage{
macument(pkg = ".", roclets = NULL, ...)
}
\arguments{
\item{pkg}{Package description, can be path or package name.  See
\code{as.package} in \code{devtools} for more information.}

\item{roclets}{Character vector of roclet names to use with package.
This defaults to \code{NULL}, which will use the \code{roclets} fields
in the list provided in the \code{Roxygen} DESCRIPTION field. If none
are specified, defaults to \code{c("collate", "namespace", "rd")}.}

\item{...}{Additional parameters are passed to
  \code{devtools::document}.}
}
\description{
Use Markdown and Roxygen to document a package
}

